\name{parasiticbehaviorsim-package}
\alias{parasiticbehaviorsim-package}
\alias{parasiticbehaviorsim} 
\docType{package}
\title{Simulation code for simple models of coevolution of social learning and parasitic behavior}
\description{
  This package contains the code used to produce the simulation results and figures from the paper ``Parasitic behavior may hurt or help social learners''.
}

\details{
  \tabular{ll}{
    Package: \tab parasiticbehaviorsim\cr
    Type: \tab Package\cr
    Version: \tab 1.00\cr
    Date: \tab 20 May 2016\cr
    License: \tab GPL-3 \cr
  }

  The function \code{\link{sim_slpb}} conducts simulations, as defined in the paper. The examples at the end of the documentation for \code{\link{sim_slpb}} show how to reproduce the figures from the paper.
  
}
\author{
  Richard McElreath
}
\references{
  McElreath (2016). Parasitic behavior may hurt or help social learners.
}

\examples{}  

\seealso{
  \code{\link{sim_slpb}}, \code{\link{plotsim}} 
} 
\keyword{parasiticbehaviorsim} 
\keyword{package} 